removed client/pubsub.rs
